Change x and y position fields to use numeric types across requests and model

+++ app/Http/Requests/Merchant/SyncDiningAreaLayoutRequest.php
@@ -18,8 +18,8 @@ class SyncDiningAreaLayoutRequest extends FormRequest
         return [
             'tables' => ['required', 'array'],
             'tables.*.layout_type' => ['required', 'string', 'max:50'],
-            'tables.*.x_position' => ['required', 'integer', 'min:0'],
-            'tables.*.y_position' => ['required', 'integer', 'min:0'],
+            'tables.*.x_position' => ['required', 'numeric', 'min:0'],
+            'tables.*.y_position' => ['required', 'numeric', 'min:0'],
             'tables.*.table_label' => ['required', 'string', 'max:20'],
             'tables.*.table_color' => ['nullable', 'string', 'max:20', 'regex:/^#(?:[0-9a-fA-F]{3}){1,2}$/'],
             'tables.*.chair_color' => ['nullable', 'string', 'max:20', 'regex:/^#(?:[0-9a-fA-F]{3}){1,2}$/'],
